Ingestion_Service/requirements.txt
kafka-python
psycopg2-binary
requests
python-dotenv
geopy
pandas